requirements.txt
flask
requests
openai
python-dotenv
flask-caching